+++ railties/test/generators/authentication_generator_test.rb
@@ -39,9 +39,7 @@ class AuthenticationGeneratorTest < Rails::Generators::TestCase
       assert_equal "  include Authentication\n", includes_line, "includes module on first line of class definition"
     end
 
-    assert_file "Gemfile" do |content|
-      assert_match(/\ngem "bcrypt"/, content)
-    end
+    assert_gem "bcrypt"
 
     assert_file "config/routes.rb" do |content|
       assert_match(/resource :session/, content)
@@ -83,9 +81,7 @@ class AuthenticationGeneratorTest < Rails::Generators::TestCase
       assert_equal "  include Authentication\n", includes_line, "includes module on first line of class definition"
     end
 
-    assert_file "Gemfile" do |content|
-      assert_match(/\ngem "bcrypt"/, content)
-    end
+    assert_gem "bcrypt"
 
     assert_file "config/routes.rb" do |content|
       assert_match(/resource :session/, content)