+++ railties/CHANGELOG.md
@@ -1,2 +1,5 @@
+*   Add `assert_gem`, `assert_gems`, `assert_no_gem`, and `assert_no_gems` to `Rails::Generators::TestCase`
+
+    *Sean Doyle*
 
 Please check [8-0-stable](https://github.com/rails/rails/blob/8-0-stable/railties/CHANGELOG.md) for previous changes.